src/FullGraph/requirements.txt
langchain
langgraph
langchain-google-genai
transformers
google-api-python-client
google-auth-oauthlib
python-docx
reportlab
pandas
geopy
pytz
dateparser
tk
numpy
timezonefinder
openpyxl
sendgrid
uvicorn
torch
pytz
dateparser
pydantic
google-auth
google-auth-httplib2
httpx
pymilvus
cnocr
slack_sdk
